Dockerize the Litmus demo App (#11)
* Dockerize the Litmus demo App

Issue #2189 Dockerize the Litmus demo repository
to reduce the prerequisites for running a
demo experiment


* added support for eksctl
updated requirements.txt with awscli
added wrapperfile to run cmd inside container
Update Makefile

+++ manage.py
@@ -115,10 +115,12 @@ def start(args):
         """
         print_color(f"Setup kind cluster", bcolors.OKBLUE)
 
-        # Installing kind cluster
-        run_shell(f"curl -Lo ./kind https://kind.sigs.k8s.io/dl/v0.7.0/kind-$(uname)-amd64")
-        run_shell(f"chmod +x ./kind")
-        run_shell(f"mv ./kind /usr/local/bin/kind")
+        # install kind if not found
+        if not os.path.isfile('/usr/local/bin/kind'):
+            # Installing kind cluster
+            run_shell(f"curl -Lo ./kind https://kind.sigs.k8s.io/dl/v0.7.0/kind-$(uname)-amd64")
+            run_shell(f"chmod +x ./kind")
+            run_shell(f"mv ./kind /usr/local/bin/kind")
 
         # Verify the kind installation
         run_shell(f"kind version")
